Move log initialization to CreateDomain
This commit moves the insertion of the first (empty) SignedMapRoot
into the Log. This initialization keeps the revision number of the
map in sync with the log index the SignedMapRoot's get inserted at.

This commit also makes CreateDomain WAIT until the log has fully
sequenced the first SMR before returning. We don't want to return
before things are fully ready.

Previously:
- This initialization was happening late on sequencer startup.
- Initialization did not wait for the log to sequence the first
  SMR, causing tests to start in an invalid state.