Use a WaitForSTH update scheme.
Rather than blindly re-sending the UpdateEntryRequest every few seconds
until the change is observed, this commit adopts a slightly more
intelligent strategy:
1. Observe current SignedLogRoot
2. Queue update request
3. Wait for a new epoch to be created
4. Query the user's account
5. Compare against the desired state.
   a. Wait some more
   b. Recreate request and try again
   c. Do nothing, it worked :-)

This strategy is more accomadating to non-determinism in the Log.
Because we block for concrete state changes, we can be confident of what
the resulting values should or should not be, making testing feisible.